Add -Wall to compile flags.

+++ robot_controllers/src/follow_joint_trajectory.cpp
@@ -217,7 +217,7 @@ void FollowJointTrajectoryController::update(const ros::Time& now, const ros::Du
       }
 
       // Fill in actual
-      for (int j = 0; j < joints_.size(); ++j)
+      for (size_t j = 0; j < joints_.size(); ++j)
       {
         feedback_.actual.positions[j] = joints_[j]->getPosition();
         feedback_.actual.velocities[j] = joints_[j]->getVelocity();
@@ -225,7 +225,7 @@ void FollowJointTrajectoryController::update(const ros::Time& now, const ros::Du
       }
 
       // Fill in error
-      for (int j = 0; j < joints_.size(); ++j)
+      for (size_t j = 0; j < joints_.size(); ++j)
       {
         feedback_.error.positions[j] = shortest_angular_distance(feedback_.desired.positions[j],
                                                                  feedback_.actual.positions[j]);
